Added auth restrictions by user type, updated domain model, updated some services.

+++ Backend/APZ-BACKEND/APZ-BACKEND.Presentation/Controllers/BusinessUsersController.cs
@@ -4,6 +4,7 @@ using APZ_BACKEND.Core.Exceptions;
 using APZ_BACKEND.Core.Helpers;
 using APZ_BACKEND.Core.Services.Users;
 using APZ_BACKEND.Core.Services.Users.BusinessUsers;
+using APZ_BACKEND.Presentation.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Options;
@@ -59,6 +60,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpGet("account-data")]
 		public async Task<IActionResult> GetAccountData()
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			int contextUserId = int.Parse(HttpContext.User.Identity.Name);
 
 			var result = await userService.GetAccountData(contextUserId);
@@ -105,6 +109,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpPut]
 		public async Task<IActionResult> Update([FromBody]UpdateBusinessUserRequest businessUser)
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			int contextUserId = int.Parse(HttpContext.User.Identity.Name);
 
 			var result = await userService.UpdateBusinessUser(businessUser, contextUserId);
@@ -117,6 +124,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpDelete]
 		public async Task<IActionResult> Delete()
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			int contextUserId = int.Parse(HttpContext.User.Identity.Name);
 
 			var result = await userService.DeleteBusinessUser(contextUserId);
@@ -134,7 +144,8 @@ namespace APZ_BACKEND.Presentation.Controllers
 			{
 				Subject = new ClaimsIdentity(new Claim[]
 				{
-					new Claim(ClaimTypes.Name, userId.ToString())
+					new Claim(ClaimTypes.Name, userId.ToString()),
+					new Claim(ClaimTypes.Role, Constants.Users.BusinessUserFlag.ToString())
 				}),
 				Expires = DateTime.UtcNow.AddDays(7),
 				SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
