+++ Backend/APZ-BACKEND/APZ-BACKEND.Presentation/Controllers/EmployeesRolesController.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using APZ_BACKEND.Core.Dtos.EmployeesRoles;
 using APZ_BACKEND.Core.Services.EmployeesRoles;
+using APZ_BACKEND.Presentation.Helpers;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
@@ -24,6 +25,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpGet("business-roles")]
 		public async Task<IActionResult> GetBusinessEmployeesRoles()
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			int contextUserId = int.Parse(HttpContext.User.Identity.Name);
 
 			var employeesRoles = await employeesRolesService.GetBusinessEmployeesRoles(contextUserId);
@@ -33,6 +37,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpPost("create-employees-role")]
 		public async Task<IActionResult> CreateEmployeesRole(CreateEmployeesRoleDto createDto)
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			int contextUserId = int.Parse(HttpContext.User.Identity.Name);
 
 			var result = await employeesRolesService.Create(contextUserId, createDto);
@@ -45,6 +52,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpPost("add-employee-to-role")]
 		public async Task<IActionResult> AddEmployeeToRole(AddEmployeeToRole addEmployeeToRoleDto)
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			var result = await employeesRolesService.AddEmployeeToRole(addEmployeeToRoleDto.EmployeeId, addEmployeeToRoleDto.RoleId);
 			if (!result.Success)
 				return BadRequest(new { message = result.ErrorMessage });
@@ -55,6 +65,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpPut("employees-role")]
 		public async Task<IActionResult> UpdateEmployeesRole(EmployeesRoleDto roleDto)
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			var result = await employeesRolesService.Update(roleDto);
 			if (!result.Success)
 				return BadRequest(new { message = result.ErrorMessage });
@@ -65,6 +78,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpDelete("employees-role/{roleId}")]
 		public async Task<IActionResult> DeleteEmployeesRole(int roleId)
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			var result = await employeesRolesService.Delete(roleId);
 			if (!result.Success)
 				return BadRequest(new { message = result.ErrorMessage });
@@ -75,6 +91,9 @@ namespace APZ_BACKEND.Presentation.Controllers
 		[HttpDelete("employee-in-role/{roleId}/{employeeId}")]
 		public async Task<IActionResult> RemoveEmployeeFromRole(int roleId, int employeeId)
 		{
+			if (!ContextAuthHelper.IsBusinessUser(HttpContext.User.Claims))
+				return BadRequest(new { message = "Current user is not a businessUser" });
+
 			var result = await employeesRolesService.RemoveEmployeeFromRole(roleId, employeeId);
 			if (!result.Success)
 				return BadRequest(new { message = result.ErrorMessage });
